chore: Refactor cast module temporal types (#3624)
