+++ native/spark-expr/src/conversion_funcs/mod.rs
@@ -19,4 +19,5 @@ mod boolean;
 pub mod cast;
 mod numeric;
 mod string;
+mod temporal;
 mod utils;
